clean-alpha-prediction-file
