+++ alpha_prediction.py
@@ -11,14 +11,13 @@ from collections import defaultdict
 from augmentation import ColourDistortion, TensorNormalise, ModuleCompose
 from models import *
 from configs import get_datasets, get_mean_std
-from evaluate import train_reg, test_reg, test_reg_component, encode_train_set_spirograph
-from tqdm import tqdm
+from evaluate import train_reg, test_reg
 
 parser = argparse.ArgumentParser(description='Tune regularization coefficient of downstream classifier.')
 parser.add_argument("--num-workers", type=int, default=2, help='Number of threads for data loaders')
-parser.add_argument("--baselines", type=str, default='ckpt', help='File series to load for baseline')
-parser.add_argument("--ours", type=str, default='invgpn', help='File series to load for our method')
-parser.add_argument("--reg", type=float, default=1e-8, help='Regularization parameter')
+parser.add_argument("--reg", type=float, default=1e-5, help='Regularization parameter')
+parser.add_argument("--load-from", type=str, default='ckpt.pth', help='File to load from')
+
 args = parser.parse_args()
 
 
@@ -34,12 +33,13 @@ def get_loss(fname):
     device = 'cuda' if torch.cuda.is_available() else 'cpu'
 
     # Data
-    trainset, testset, clftrainset, num_classes, stem, col_distort, batch_transform = get_datasets(args.dataset, train_proportion=1.0)
+    _, testset, clftrainset, _, stem, col_distort, batch_transform = get_datasets(
+        args.dataset, augment_clf_train=True, augment_test=True)
 
     testloader = torch.utils.data.DataLoader(testset, batch_size=1000, shuffle=False, num_workers=args.num_workers,
                                              pin_memory=True)
     clftrainloader = torch.utils.data.DataLoader(clftrainset, batch_size=1000, shuffle=False, num_workers=args.num_workers,
-                                             pin_memory=True)
+                                                 pin_memory=True)
 
     # Model
     ##############################################################
@@ -65,22 +65,38 @@ def get_loss(fname):
 
     batch_transform = batch_transform.to(device)
 
-    X, y = encode_train_set_spirograph(clftrainloader, device, net, col_distort, batch_transform)
-    X_test, y_test = encode_train_set_spirograph(testloader, device, net, col_distort, batch_transform)
+
+    def create_dataset(clftrainloader, device, net, target=None):
+        if target is None:
+            target = device
+
+        net.eval()
+
+        with torch.no_grad():
+            store = []
+            for batch_idx, (inputs, _) in enumerate(clftrainloader):
+                inputs = inputs.to(device)
+                rn = col_distort.sample_random_numbers(inputs.shape, inputs.device)
+                inputs = batch_transform(inputs, rn)
+                representation = net(inputs)
+                representation = representation.to(target)
+                store.append((representation, rn))
+
+            Xi, y = zip(*store)
+            Xi, y = torch.cat(Xi, dim=0), torch.cat(y, dim=0)
+
+        return Xi, y
+        
+
+    X, y = create_dataset(clftrainloader, device, net)
+    X_test, y_test = create_dataset(testloader, device, net)
     clf = train_reg(X, y, device, reg_weight=args.reg)
-    loss = test_reg_component(X_test, y_test, clf)
+    loss = test_reg(X_test, y_test, clf)
+    print('Loss for augmentation parameter prediction')
+    print(loss)
     return loss
 
 
 
-baselines = args.baselines.split(",")
-ours = args.ours.split(",")
-results = defaultdict(list)
-for stem in baselines+ours:
-    for epoch in range(49, 50, 10):
-        print(epoch)
-        fname = stem + '_epoch{:03d}.pth'.format(epoch)
-        loss = get_loss(fname)
-        results[stem].append(loss)
+loss = get_loss(args.load_from)
 
-print(results)
